Moving everything out of the __init__ file.
